requirements.txt
groq==0.28.0
pdf2image==1.17.0
pip-chill==1.0.3
pymupdf==1.26.1
pytesseract==0.3.13
python-dotenv==1.1.0
rapidfuzz==3.13.0